refactor(commands): extracted alias commands
